+++ compose.yml
@@ -73,7 +73,6 @@ services:
       - node_modules_ghost_minifier:/home/ghost/ghost/minifier/node_modules:delegated
       - node_modules_ghost_model-to-domain-event-interceptor:/home/ghost/ghost/model-to-domain-event-interceptor/node_modules:delegated
       - node_modules_ghost_mw-api-version-mismatch:/home/ghost/ghost/mw-api-version-mismatch/node_modules:delegated
-      - node_modules_ghost_mw-cache-control:/home/ghost/ghost/mw-cache-control/node_modules:delegated
       - node_modules_ghost_mw-error-handler:/home/ghost/ghost/mw-error-handler/node_modules:delegated
       - node_modules_ghost_mw-version-match:/home/ghost/ghost/mw-version-match/node_modules:delegated
       - node_modules_ghost_mw-vhost:/home/ghost/ghost/mw-vhost/node_modules:delegated
@@ -224,7 +223,6 @@ volumes:
   node_modules_ghost_minifier: {}
   node_modules_ghost_model-to-domain-event-interceptor: {}
   node_modules_ghost_mw-api-version-mismatch: {}
-  node_modules_ghost_mw-cache-control: {}
   node_modules_ghost_mw-error-handler: {}
   node_modules_ghost_mw-version-match: {}
   node_modules_ghost_mw-vhost: {}
