+++ ghost/core/core/server/web/shared/middleware/cache-control.js
