--- ghost/mw-cache-control/README.md
@@ -1,23 +1,0 @@
-# Mw Cache Control
-
-Cache-Control middleware for Ghost
-
-
-## Usage
-
-
-## Develop
-
-This is a monorepo package.
-
-Follow the instructions for the top-level repo.
-1. `git clone` this repo & `cd` into it as usual
-2. Run `yarn` to install top-level dependencies.
-
-
-
-## Test
-
-- `yarn lint` run just eslint
-- `yarn test` run lint and tests
-
