--- ghost/mw-cache-control/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/mw-cache-control');
